# src/main/java/com/tallyo/tallyo_backend/dto/ApiError.java
package com.tallyo.tallyo_backend.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;

@Getter
@Builder
@AllArgsConstructor
public class ApiError {
    private String code;
    private String message;
    private String details;
    private String path;
    private String timestamp;
}